requirements.txt
pytest>=7.0
requests>=2.31
pyyaml>=6.0
pandas>=2.0
openpyxl>=3.1
allure-pytest>=2.13
